+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Meer data toegevoegd aan prison leave log
 - General statebag handler wrappers added to avoid code duplication
 - Tankstations + garages voor helikopter/vliegtuig toegevoegd
+- Je kan nu je armor terug afdoen via F1 of het `/retrieveArmor` command
 
 ### Changed